+++ src/store/getters/getters.js
@@ -1,0 +1,8 @@
+export default {
+  PRODUCTS(state) {
+    return state.products;
+  },
+  CART(state) {
+    return state.cart;
+  }
+};
